refactor select query generation
-> Improve 'Aggregate' query generation logic
-> Move query generation code from 'Select.hs' to 'Select/Internal.hs'
